absctract task excutor
