+++ packages/app/android/src/main/java/io/invertase/firebase/common/TaskExecutorService.java
@@ -1,0 +1,108 @@
+package io.invertase.firebase.common;
+
+/*
+ * Copyright (c) 2016-present Invertase Limited & Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this library except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Set;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.RejectedExecutionHandler;
+import java.util.concurrent.ThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.SynchronousQueue;
+
+public class TaskExecutorService {
+  private final String name;
+  private final int maximumPoolSize;
+  private final int keepAliveSeconds;
+  private static Map<String, ExecutorService> executors = new HashMap<>();
+
+  TaskExecutorService(
+    String name,
+    int maximumPoolSize,
+    int keepAliveSeconds
+  ) {
+    this.name = name;
+    this.maximumPoolSize = maximumPoolSize;
+    this.keepAliveSeconds = keepAliveSeconds;
+  }
+
+  public ExecutorService getExecutor() {
+    return getExecutor(false, "");
+  }
+
+  public ExecutorService getTransactionalExecutor() {
+    return getExecutor(true, "");
+  }
+
+  public ExecutorService getTransactionalExecutor(String identifier) {
+    return getExecutor(true, identifier);
+  }
+
+  public ExecutorService getExecutor(boolean isTransactional, String identifier) {
+    String executorName = getExecutorName(isTransactional, identifier);
+    ExecutorService existingExecutor = executors.get(executorName);
+    if (existingExecutor != null) return existingExecutor;
+    ExecutorService newExecutor = getNewExecutor(isTransactional);
+    executors.put(executorName, newExecutor);
+    return newExecutor;
+  }
+
+  private ExecutorService getNewExecutor(boolean isTransactional) {
+    if (isTransactional == true) {
+      return Executors.newSingleThreadExecutor();
+    } else {
+      ThreadPoolExecutor threadPoolExecutor = new ThreadPoolExecutor(0, maximumPoolSize, keepAliveSeconds, TimeUnit.SECONDS, new SynchronousQueue<Runnable>());
+      threadPoolExecutor.setRejectedExecutionHandler(executeInFallback);
+      return threadPoolExecutor;
+    }
+  }
+
+  private RejectedExecutionHandler executeInFallback = new RejectedExecutionHandler() {
+    public void rejectedExecution(Runnable r, ThreadPoolExecutor executor) {
+      ExecutorService fallbackExecutor = getTransactionalExecutor();
+      fallbackExecutor.execute(r);
+    };
+  };
+
+  public String getExecutorName(boolean isTransactional, String identifier) {
+    if (isTransactional == true) {
+      return name + "TransactionalExecutor" + identifier;
+    }
+    return name + "Executor" + identifier;
+  }
+
+  public void shutdown() {
+    Set<String> existingExecutorNames = executors.keySet();
+    existingExecutorNames.removeIf((executorName) -> {
+      return executorName.startsWith(name) == false;
+    });
+    existingExecutorNames.forEach((executorName) -> {
+      removeExecutor(executorName);
+    });
+  }
+
+  public void removeExecutor(String executorName) {
+    ExecutorService existingExecutor = executors.get(executorName);
+    if (existingExecutor != null) {
+      existingExecutor.shutdownNow();
+      executors.remove(executorName);
+    }
+  }
+}
